refine uses multiple dispatch
