+++ sympy/__init__.py
@@ -47,6 +47,7 @@ from .integrals import *
 from .tensor import *
 from .parsing import *
 from .calculus import *
+from .assumptions import refine_handlers
 # Adds about .04-.05 seconds of import time
 # from combinatorics import *
 # This module is slow to import:
